Extract Lookup component
